Refactor pkg names
