--- commands.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/bwmarrin/discordgo"
-)
-
-type CustomCommandDef interface {
-	Definition() *discordgo.ApplicationCommand
-	Handler(sess *discordgo.Session, inter *discordgo.InteractionCreate) error
-}
