+++ internal/command/play.go
@@ -1,4 +1,4 @@
-package commands
+package command
 
 import (
 	"sync"
